+++ src/main/java/com/ghostipedia/cosmiccore/api/capability/recipe/SoulRecipeCapability.java
@@ -31,7 +31,7 @@ public class SoulRecipeCapability extends RecipeCapability<Integer> {
 
     @Override
     public Integer copyWithModifier(Integer content, ContentModifier modifier) {
-        return modifier.apply(content).intValue();
+        return modifier.apply(content);
     }
 
     @Override
